+++ scripts/subclass/experimental-elixir-brew.js
@@ -1,17 +1,29 @@
-const version = "14.5.0";
+const version = "14.5.1";
 const optionName = "Experimental Elixir - Brew";
 const elixirId = "Compendium.fvtt-trazzm-homebrew-5e.trazzm-automation-items-2024.Item.3HYJK61ZGLlQKfi4";
+const _flagGroup = "fvtt-trazzm-homebrew-5e";
+const _flagName = "experimental-elixir-choice";
 
 try {
     if (args[0].macroPass === "postActiveEffects") {
         let elixirItem = await fromUuid(elixirId);
         const artificerLevel = actor.getRollData().classes?.artificer?.levels ?? 1;
         const intMod = actor.system.abilities.int.mod;
+        let elixirCount = Number(actor.system.scale.alchemist['experimental-elixir-count']);
+
+        const elixirDupe = foundry.utils.duplicate(elixirItem);
+
+        // should the user be allowed to select the effect?
+        if (workflow.activity.identifier === 'spell-slot-for-elixir') {
+            elixirCount = 1;
+        }
+        elixirDupe.system.quantity = elixirCount;
 
         // add the food to the actors inventory, they can dispense it
-        let tempItem = elixirItem.toObject();
-        tempItem.system.quantity = 2;
-        let results = await actor.createEmbeddedDocuments('Item',[tempItem]);
+        let results = await actor.createEmbeddedDocuments('Item',[elixirDupe]);
+        if (workflow.activity.identifier === 'spell-slot-for-elixir') {
+            await results[0].setFlag(_flagGroup, _flagName, true);
+        }
 
         // update the activity effects
         let healingActivity = results[0].system.activities.getName("Healing");